+++ tests/oidc_resource_server/test_clients.py
@@ -8,58 +8,95 @@ import pytest
 from joserfc.jwk import KeySet, RSAKey
 from requests.exceptions import HTTPError
 
-from lasuite.oidc_resource_server.clients import AuthorizationServerClient
+from lasuite.oidc_resource_server.clients import AuthorizationServerClient, JWTAuthorizationServerClient
 
 
-@pytest.fixture(name="client")
-def fixture_client():
+@pytest.fixture(name="authorization_server_client")
+def fixture_authorization_server_client(settings):
     """Generate an Authorization Server client."""
-    return AuthorizationServerClient(
-        url="https://auth.example.com/api/v2",
-        url_jwks="https://auth.example.com/api/v2/jwks",
-        url_introspection="https://auth.example.com/api/v2/introspect",
-        verify_ssl=True,
-        timeout=5,
-        proxy=None,
-    )
+    settings.OIDC_OP_URL = "https://auth.example.com/api/v2"
+    settings.OIDC_VERIFY_SSL = True
+    settings.OIDC_TIMEOUT = 5
+    settings.OIDC_PROXY = None
+    settings.OIDC_OP_INTROSPECTION_ENDPOINT = "https://auth.example.com/api/v2/introspect"
+
+    return AuthorizationServerClient()
 
 
-def test_authorization_server_client_initialization():
+@pytest.fixture(name="jwt_authorization_server_client")
+def fixture_jwt_authorization_server_client(settings):
+    """Generate an Authorization Server client using JWT."""
+    settings.OIDC_OP_URL = "https://auth.example.com/api/v2"
+    settings.OIDC_VERIFY_SSL = True
+    settings.OIDC_TIMEOUT = 5
+    settings.OIDC_PROXY = None
+    settings.OIDC_OP_INTROSPECTION_ENDPOINT = "https://auth.example.com/api/v2/introspect"
+    settings.OIDC_OP_JWKS_ENDPOINT = "https://auth.example.com/api/v2/jwks"
+
+    return JWTAuthorizationServerClient()
+
+
+def test_authorization_server_client_initialization(settings):
     """Test the AuthorizationServerClient initialization."""
-    new_client = AuthorizationServerClient(
-        url="https://auth.example.com/api/v2",
-        url_jwks="https://auth.example.com/api/v2/jwks",
-        url_introspection="https://auth.example.com/api/v2/checktoken/foo",
-        verify_ssl=True,
-        timeout=5,
-        proxy=None,
-    )
+    settings.OIDC_OP_URL = "https://auth.example.com/api/v2"
+    settings.OIDC_VERIFY_SSL = True
+    settings.OIDC_TIMEOUT = 5
+    settings.OIDC_PROXY = None
+    settings.OIDC_OP_INTROSPECTION_ENDPOINT = "https://auth.example.com/api/v2/introspect"
+
+    new_client = AuthorizationServerClient()
 
     assert new_client.url == "https://auth.example.com/api/v2"
-    assert new_client._url_introspection == "https://auth.example.com/api/v2/checktoken/foo"
+    assert new_client._url_introspection == "https://auth.example.com/api/v2/introspect"
+    assert new_client._verify_ssl is True
+    assert new_client._timeout == 5
+    assert new_client._proxy is None
+
+
+def test_jwt_authorization_server_client_initialization(settings):
+    """Test the JWTAuthorizationServerClient initialization."""
+    settings.OIDC_OP_URL = "https://auth.example.com/api/v2"
+    settings.OIDC_VERIFY_SSL = True
+    settings.OIDC_TIMEOUT = 5
+    settings.OIDC_PROXY = None
+    settings.OIDC_OP_INTROSPECTION_ENDPOINT = "https://auth.example.com/api/v2/introspect"
+    settings.OIDC_OP_JWKS_ENDPOINT = "https://auth.example.com/api/v2/jwks"
+
+    new_client = JWTAuthorizationServerClient()
+
+    assert new_client.url == "https://auth.example.com/api/v2"
+    assert new_client._url_introspection == "https://auth.example.com/api/v2/introspect"
     assert new_client._url_jwks == "https://auth.example.com/api/v2/jwks"
     assert new_client._verify_ssl is True
     assert new_client._timeout == 5
     assert new_client._proxy is None
 
 
-def test_introspection_headers(client):
+def test_introspection_headers(authorization_server_client):
     """Test the introspection headers to ensure they match the expected values."""
-    assert client._introspection_headers == {
+    assert authorization_server_client._introspection_headers == {
+        "Content-Type": "application/x-www-form-urlencoded",
+        "Accept": "application/json",
+    }
+
+
+def test_jwt_introspection_headers(jwt_authorization_server_client):
+    """Test the introspection headers to ensure they match the expected values."""
+    assert jwt_authorization_server_client._introspection_headers == {
         "Content-Type": "application/x-www-form-urlencoded",
         "Accept": "application/token-introspection+jwt",
     }
 
 
 @patch("requests.post")
-def test_get_introspection_success(mock_post, client):
+def test_get_introspection_success(mock_post, authorization_server_client):
     """Test 'get_introspection' method with a successful response."""
     mock_response = MagicMock()
     mock_response.raise_for_status.return_value = None
     mock_response.text = "introspection response"
     mock_post.return_value = mock_response
 
-    result = client.get_introspection("client_id", "client_secret", "token")
+    result = authorization_server_client.get_introspection("client_id", "client_secret", "token")
     assert result == "introspection response"
 
     mock_post.assert_called_once_with(
@@ -71,7 +108,7 @@ def test_get_introspection_success(mock_post, client):
         },
         headers={
             "Content-Type": "application/x-www-form-urlencoded",
-            "Accept": "application/token-introspection+jwt",
+            "Accept": "application/json",
         },
         verify=True,
         timeout=5,
@@ -81,44 +118,44 @@ def test_get_introspection_success(mock_post, client):
 
 @patch("requests.post", side_effect=HTTPError())
 # pylint: disable=(unused-argument
-def test_get_introspection_error(mock_post, client):
+def test_get_introspection_error(mock_post, authorization_server_client):
     """Test 'get_introspection' method with an HTTPError."""
     with pytest.raises(HTTPError):
-        client.get_introspection("client_id", "client_secret", "token")
+        authorization_server_client.get_introspection("client_id", "client_secret", "token")
 
 
 @patch("requests.get")
-def test_get_jwks_success(mock_get, client):
+def test_get_jwks_success(mock_get, jwt_authorization_server_client):
     """Test 'get_jwks' method with a successful response."""
     mock_response = MagicMock()
     mock_response.raise_for_status.return_value = None
     mock_response.json.return_value = {"jwks": "foo"}
     mock_get.return_value = mock_response
 
-    result = client.get_jwks()
+    result = jwt_authorization_server_client.get_jwks()
     assert result == {"jwks": "foo"}
 
     mock_get.assert_called_once_with(
         "https://auth.example.com/api/v2/jwks",
-        verify=client._verify_ssl,
-        timeout=client._timeout,
-        proxies=client._proxy,
+        verify=jwt_authorization_server_client._verify_ssl,
+        timeout=jwt_authorization_server_client._timeout,
+        proxies=jwt_authorization_server_client._proxy,
     )
 
 
 @patch("requests.get")
-def test_get_jwks_error(mock_get, client):
+def test_get_jwks_error(mock_get, jwt_authorization_server_client):
     """Test 'get_jwks' method with an HTTPError."""
     mock_response = MagicMock()
     mock_response.raise_for_status.side_effect = HTTPError(response=MagicMock(status=500))
     mock_get.return_value = mock_response
 
     with pytest.raises(HTTPError):
-        client.get_jwks()
+        jwt_authorization_server_client.get_jwks()
 
 
 @patch("requests.get")
-def test_import_public_keys_valid(mock_get, client):
+def test_import_public_keys_valid(mock_get, jwt_authorization_server_client):
     """Test 'import_public_keys' method with a successful response."""
     mocked_key = RSAKey.generate_key(2048)
 
@@ -127,39 +164,39 @@ def test_import_public_keys_valid(mock_get, client):
     mock_response.json.return_value = {"keys": [mocked_key.as_dict()]}
     mock_get.return_value = mock_response
 
-    response = client.import_public_keys()
+    response = jwt_authorization_server_client.import_public_keys()
 
     assert isinstance(response, KeySet)
     assert response.as_dict() == KeySet([mocked_key]).as_dict()
 
 
 @patch("requests.get")
-def test_import_public_keys_http_error(mock_get, client):
+def test_import_public_keys_http_error(mock_get, jwt_authorization_server_client):
     """Test 'import_public_keys' method with an HTTPError."""
     mock_response = MagicMock()
     mock_response.raise_for_status.side_effect = HTTPError(response=MagicMock(status=500))
     mock_get.return_value = mock_response
 
     with pytest.raises(HTTPError):
-        client.import_public_keys()
+        jwt_authorization_server_client.import_public_keys()
 
 
 @patch("requests.get")
-def test_import_public_keys_empty_jwks(mock_get, client):
+def test_import_public_keys_empty_jwks(mock_get, jwt_authorization_server_client):
     """Test 'import_public_keys' method with empty keys response."""
     mock_response = MagicMock()
     mock_response.raise_for_status.return_value = None
     mock_response.json.return_value = {"keys": []}
     mock_get.return_value = mock_response
 
-    response = client.import_public_keys()
+    response = jwt_authorization_server_client.import_public_keys()
 
     assert isinstance(response, KeySet)
     assert response.as_dict() == {"keys": []}
 
 
 @patch("requests.get")
-def test_import_public_keys_invalid_jwks(mock_get, client):
+def test_import_public_keys_invalid_jwks(mock_get, jwt_authorization_server_client):
     """Test 'import_public_keys' method with invalid keys response."""
     mock_response = MagicMock()
     mock_response.raise_for_status.return_value = None
@@ -167,4 +204,4 @@ def test_import_public_keys_invalid_jwks(mock_get, client):
     mock_get.return_value = mock_response
 
     with pytest.raises(ValueError, match="Missing key type"):
-        client.import_public_keys()
+        jwt_authorization_server_client.import_public_keys()